refactor: Make bot stupid (databaseless)

+++ server/src/telegram/services/Telegram.service.ts
@@ -4,7 +4,7 @@ import { TelegramEntity } from '../entities/Telegram.entity';
 import { Repository } from 'typeorm';
 import { UserService } from '../../user/User.service';
 import { NotFoundException } from '@nestjs/common';
-import { UserEntity } from '../../user/entities/User.entity';
+import { request } from 'http';
 
 interface ITelegramService {
   create(data: CreateTelegramRequestDto): Promise<void>;
@@ -20,7 +20,9 @@ export class TelegramService implements ITelegramService {
   ) {}
 
   async create(data: CreateTelegramRequestDto): Promise<void> {
-    await this.repository.save({ ...data });
+    const user = await this.userService.findById(data.userId);
+    if (!user) throw new NotFoundException('User not found');
+    await this.repository.save({ ...data, user });
   }
 
   async accept(id: string): Promise<void> {
@@ -31,16 +33,40 @@ export class TelegramService implements ITelegramService {
       },
     });
     if (!request) throw new NotFoundException('Telegram request not found');
+    const user = request.user;
+    if (!user) throw new NotFoundException('User not found');
+    user.telegramId = request.telegramId;
 
-    await this.userService.changeUser({
-      ...request.user,
-      telegramId: request.telegramId,
-    });
-
+    await this.userService.changeUser(user);
     await this.repository.delete(id);
+
+    await fetch(process.env.BOT_URL + '/review_result', {
+      method: 'POST',
+      body: JSON.stringify({
+        chat_id: request.chatId,
+        lang: request.lang,
+        result: true,
+      }),
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    });
   }
 
   async deny(id: string): Promise<void> {
-    await this.repository.delete(id);
+    const request = await this.repository.findOneBy({ id });
+    if (!request) throw new NotFoundException('Request not found');
+    await this.repository.delete(request.id);
+    await fetch(process.env.BOT_URL + '/review_result', {
+      method: 'POST',
+      body: JSON.stringify({
+        chat_id: request.chatId,
+        lang: request.lang,
+        result: false,
+      }),
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    });
   }
 }
